@@ -1,28 +1,43 @@
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
+
+def getGitSha = { ->
+    def stdout = new ByteArrayOutputStream()
+    exec {
+        commandLine 'git', 'rev-parse', '--short' , 'HEAD'
+        standardOutput = stdout
+    }
+    return stdout.toString().trim()
+}
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.keylesspalace.tusky"
-        minSdkVersion 18
-        targetSdkVersion 27
-        versionCode 34
-        versionName "1.5.2"
+        minSdkVersion 21
+        targetSdkVersion 28
+        versionCode 49
+        versionName "3.1"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        vectorDrawables.useSupportLibrary true
+        vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
         release {
             minifyEnabled true
             shrinkResources true
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            proguardFiles 'proguard-rules.pro'
         }
-        debug {
+        debug { }
+    }
+
+    flavorDimensions "color"
+    productFlavors {
+        blue { }
+        green {
             applicationIdSuffix ".test"
+            versionNameSuffix "-"+getGitSha()
         }
     }
 
