@@ -1,31 +1,32 @@
 package com.keylesspalace.tusky.adapter;
 
+import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.keylesspalace.tusky.R;
 import com.keylesspalace.tusky.entity.Account;
 import com.keylesspalace.tusky.interfaces.AccountActionListener;
-import com.pkmmte.view.CircularImageView;
+import com.keylesspalace.tusky.util.CustomEmojiHelper;
 import com.squareup.picasso.Picasso;
 
 public class MutesAdapter extends AccountAdapter {
-    private static final int VIEW_TYPE_MUTED_USER = 0;
-    private static final int VIEW_TYPE_FOOTER = 1;
 
     public MutesAdapter(AccountActionListener accountActionListener) {
         super(accountActionListener);
     }
 
+    @NonNull
     @Override
-    public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         switch (viewType) {
             default:
-            case VIEW_TYPE_MUTED_USER: {
+            case VIEW_TYPE_ACCOUNT: {
                 View view = LayoutInflater.from(parent.getContext())
                         .inflate(R.layout.item_muted_user, parent, false);
                 return new MutesAdapter.MutedUserViewHolder(view);
@@ -33,34 +34,23 @@ public class MutesAdapter extends AccountAdapter {
             case VIEW_TYPE_FOOTER: {
                 View view = LayoutInflater.from(parent.getContext())
                         .inflate(R.layout.item_footer, parent, false);
-                return new FooterViewHolder(view);
+                return new LoadingFooterViewHolder(view);
             }
         }
     }
 
     @Override
-    public void onBindViewHolder(RecyclerView.ViewHolder viewHolder, int position) {
-        if (position < accountList.size()) {
+    public void onBindViewHolder(@NonNull RecyclerView.ViewHolder viewHolder, int position) {
+        if (getItemViewType(position) == VIEW_TYPE_ACCOUNT) {
             MutedUserViewHolder holder = (MutedUserViewHolder) viewHolder;
             holder.setupWithAccount(accountList.get(position));
-            holder.setupActionListener(accountActionListener, true, position);
-        } else {
-            FooterViewHolder holder = (FooterViewHolder) viewHolder;
-            holder.setState(footerState);
+            holder.setupActionListener(accountActionListener);
         }
     }
 
-    @Override
-    public int getItemViewType(int position) {
-        if (position == accountList.size()) {
-            return VIEW_TYPE_FOOTER;
-        } else {
-            return VIEW_TYPE_MUTED_USER;
-        }
-    }
 
     static class MutedUserViewHolder extends RecyclerView.ViewHolder {
-        private CircularImageView avatar;
+        private ImageView avatar;
         private TextView username;
         private TextView displayName;
         private ImageButton unmute;
@@ -76,7 +66,8 @@ public class MutesAdapter extends AccountAdapter {
 
         void setupWithAccount(Account account) {
             id = account.getId();
-            displayName.setText(account.getName());
+            CharSequence emojifiedName = CustomEmojiHelper.emojifyString(account.getName(), account.getEmojis(), displayName);
+            displayName.setText(emojifiedName);
             String format = username.getContext().getString(R.string.status_username_format);
             String formattedUsername = String.format(format, account.getUsername());
             username.setText(formattedUsername);
@@ -86,20 +77,9 @@ public class MutesAdapter extends AccountAdapter {
                     .into(avatar);
         }
 
-        void setupActionListener(final AccountActionListener listener, final boolean muted,
-                final int position) {
-            unmute.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    listener.onMute(!muted, id, position);
-                }
-            });
-            avatar.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    listener.onViewAccount(id);
-                }
-            });
+        void setupActionListener(final AccountActionListener listener) {
+            unmute.setOnClickListener(v -> listener.onMute(false, id, getAdapterPosition()));
+            avatar.setOnClickListener(v -> listener.onViewAccount(id));
         }
     }
 }
