@@ -32,21 +32,26 @@ import android.view.View
 import android.widget.EditText
 import android.widget.TextView
 import com.keylesspalace.tusky.db.AccountManager
+import com.keylesspalace.tusky.di.Injectable
 import com.keylesspalace.tusky.entity.AccessToken
 import com.keylesspalace.tusky.entity.AppCredentials
 import com.keylesspalace.tusky.network.MastodonApi
 import com.keylesspalace.tusky.util.CustomTabsHelper
-import com.keylesspalace.tusky.util.OkHttpUtils
 import com.keylesspalace.tusky.util.ThemeUtils
 import kotlinx.android.synthetic.main.activity_login.*
+import okhttp3.HttpUrl
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Inject
 
 
-class LoginActivity : AppCompatActivity() {
+class LoginActivity : AppCompatActivity(), Injectable {
+
+    @Inject
+    lateinit var mastodonApi: MastodonApi
+    @Inject
+    lateinit var accountManager: AccountManager
 
     private lateinit var preferences: SharedPreferences
     private var domain: String = ""
@@ -64,13 +69,16 @@ class LoginActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
 
         preferences = PreferenceManager.getDefaultSharedPreferences(this)
-        val theme = preferences.getString("appTheme", TuskyApplication.APP_THEME_DEFAULT)
-        ThemeUtils.setAppNightMode(theme)
+        val theme = preferences.getString("appTheme", ThemeUtils.APP_THEME_DEFAULT)
+        if (theme == "black") {
+            setTheme(R.style.TuskyBlackTheme)
+        }
+        ThemeUtils.setAppNightMode(theme, this)
 
         setContentView(R.layout.activity_login)
 
         if (savedInstanceState != null) {
-            domain = savedInstanceState.getString(DOMAIN)
+            domain = savedInstanceState.getString(DOMAIN)!!
             clientId = savedInstanceState.getString(CLIENT_ID)
             clientSecret = savedInstanceState.getString(CLIENT_SECRET)
         }
@@ -99,6 +107,13 @@ class LoginActivity : AppCompatActivity() {
 
     }
 
+    override fun finish() {
+        super.finish()
+        if(isAdditionalLogin()) {
+            overridePendingTransition(R.anim.slide_from_left, R.anim.slide_to_right)
+        }
+    }
+
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         if (item.itemId == android.R.id.home) {
             onBackPressed()
@@ -114,16 +129,6 @@ class LoginActivity : AppCompatActivity() {
         super.onSaveInstanceState(outState)
     }
 
-    private fun getApiFor(domain: String): MastodonApi {
-        val retrofit = Retrofit.Builder()
-                .baseUrl("https://" + domain)
-                .client(OkHttpUtils.getCompatibleClient(preferences))
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-
-        return retrofit.create(MastodonApi::class.java)
-    }
-
     /**
      * Obtain the oauth client credentials for this app. This is only necessary the first time the
      * app is run on a given server instance. So, after the first authentication, they are
@@ -133,7 +138,15 @@ class LoginActivity : AppCompatActivity() {
 
         loginButton.isEnabled = false
 
-        domain = validateDomain(domainEditText.text.toString())
+        domain = canonicalizeDomain(domainEditText.text.toString())
+
+        try {
+            HttpUrl.Builder().host(domain).scheme("https").build()
+        } catch (e: IllegalArgumentException) {
+            setLoading(false)
+            domainEditText.error = getString(R.string.error_invalid_domain)
+            return
+        }
 
         val callback = object : Callback<AppCredentials> {
             override fun onResponse(call: Call<AppCredentials>,
@@ -141,6 +154,7 @@ class LoginActivity : AppCompatActivity() {
                 if (!response.isSuccessful) {
                     loginButton.isEnabled = true
                     domainEditText.error = getString(R.string.error_failed_app_registration)
+                    setLoading(false)
                     Log.e(TAG, "App authentication failed. " + response.message())
                     return
                 }
@@ -159,16 +173,11 @@ class LoginActivity : AppCompatActivity() {
             }
         }
 
-        try {
-            getApiFor(domain)
-                    .authenticateApp(getString(R.string.app_name), oauthRedirectUri,
-                            OAUTH_SCOPES, getString(R.string.app_website))
-                    .enqueue(callback)
-            setLoading(true)
-        } catch (e: IllegalArgumentException) {
-            setLoading(false)
-            domainEditText.error = getString(R.string.error_invalid_domain)
-        }
+        mastodonApi
+                .authenticateApp(domain, getString(R.string.app_name), oauthRedirectUri,
+                        OAUTH_SCOPES, getString(R.string.app_website))
+                .enqueue(callback)
+        setLoading(true)
 
     }
 
@@ -216,7 +225,7 @@ class LoginActivity : AppCompatActivity() {
             val code = uri.getQueryParameter("code")
             val error = uri.getQueryParameter("error")
 
-            domain = preferences.getString(DOMAIN, "")
+            domain = preferences.getString(DOMAIN, "")!!
 
             if (code != null && domain.isNotEmpty()) {
                 /* During the redirect roundtrip this Activity usually dies, which wipes out the
@@ -250,7 +259,7 @@ class LoginActivity : AppCompatActivity() {
                     }
                 }
 
-                getApiFor(domain).fetchOAuthToken(clientId, clientSecret, redirectUri, code,
+                mastodonApi.fetchOAuthToken(domain, clientId, clientSecret, redirectUri, code,
                         "authorization_code").enqueue(callback)
             } else if (error != null) {
                 /* Authorization failed. Put the error response where the user can read it and they
@@ -290,14 +299,13 @@ class LoginActivity : AppCompatActivity() {
 
         setLoading(true)
 
-        TuskyApplication.getInstance(this).serviceLocator
-                .get(AccountManager::class.java)
-                .addAccount(accessToken, domain)
+        accountManager.addAccount(accessToken, domain)
 
         val intent = Intent(this, MainActivity::class.java)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
         startActivity(intent)
         finish()
+        overridePendingTransition(R.anim.explode, R.anim.explode)
     }
 
     companion object {
@@ -316,7 +324,7 @@ class LoginActivity : AppCompatActivity() {
         }
 
         /** Make sure the user-entered text is just a fully-qualified domain name.  */
-        private fun validateDomain(domain: String): String {
+        private fun canonicalizeDomain(domain: String): String {
             // Strip any schemes out.
             var s = domain.replaceFirst("http://", "")
             s = s.replaceFirst("https://", "")
