@@ -189,7 +189,7 @@ public class ReportActivity extends BaseActivity implements Injectable {
                 onFetchStatusesFailure((Exception) t);
             }
         };
-        mastodonApi.accountStatuses(accountId, null, null, null, null)
+        mastodonApi.accountStatuses(accountId, null, null, null, null, null)
                 .enqueue(callback);
     }
 
