@@ -15,6 +15,8 @@
 
 package com.keylesspalace.tusky;
 
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
@@ -33,10 +35,29 @@ import dagger.android.AndroidInjector;
 import dagger.android.DispatchingAndroidInjector;
 import dagger.android.support.HasSupportFragmentInjector;
 
-public class ViewThreadActivity extends BaseActivity implements HasSupportFragmentInjector {
+public class ViewThreadActivity extends BottomSheetActivity implements HasSupportFragmentInjector {
+
+    public static final int REVEAL_BUTTON_HIDDEN = 1;
+    public static final int REVEAL_BUTTON_REVEAL = 2;
+    public static final int REVEAL_BUTTON_HIDE = 3;
+
+    public static Intent startIntent(Context context, String id, String url) {
+        Intent intent = new Intent(context, ViewThreadActivity.class);
+        intent.putExtra(ID_EXTRA, id);
+        intent.putExtra(URL_EXTRA, url);
+        return intent;
+    }
+
+    private static final String ID_EXTRA = "id";
+    private static final String URL_EXTRA = "url";
+    private static final String FRAGMENT_TAG = "ViewThreadFragment_";
+
+    private int revealButtonState = REVEAL_BUTTON_HIDDEN;
 
     @Inject
     public DispatchingAndroidInjector<Fragment> dispatchingAndroidInjector;
+
+    private ViewThreadFragment fragment;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -52,17 +73,39 @@ public class ViewThreadActivity extends BaseActivity implements HasSupportFragme
             actionBar.setDisplayShowHomeEnabled(true);
         }
 
-        String id = getIntent().getStringExtra("id");
+        String id = getIntent().getStringExtra(ID_EXTRA);
+
+        fragment = (ViewThreadFragment)getSupportFragmentManager().findFragmentByTag(FRAGMENT_TAG + id);
+        if(fragment == null) {
+            fragment = ViewThreadFragment.newInstance(id);
+        }
+
         FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
-        Fragment fragment = ViewThreadFragment.newInstance(id);
-        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        fragmentTransaction.replace(R.id.fragment_container, fragment, FRAGMENT_TAG + id);
         fragmentTransaction.commit();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.view_thread_toolbar, menu);
+        MenuItem menuItem = menu.findItem(R.id.action_reveal);
+        menuItem.setVisible(revealButtonState != REVEAL_BUTTON_HIDDEN);
+        menuItem.setIcon(revealButtonState == REVEAL_BUTTON_REVEAL ?
+        R.drawable.ic_eye_24dp : R.drawable.ic_hide_media_24dp);
         return super.onCreateOptionsMenu(menu);
+    }
+
+    public void setRevealButtonState(int state) {
+        switch (state) {
+            case REVEAL_BUTTON_HIDDEN:
+            case REVEAL_BUTTON_REVEAL:
+            case REVEAL_BUTTON_HIDE:
+                this.revealButtonState = state;
+                invalidateOptionsMenu();
+                break;
+            default:
+                throw new IllegalArgumentException("Invalid reveal button state: " + state);
+        }
     }
 
     @Override
@@ -73,7 +116,11 @@ public class ViewThreadActivity extends BaseActivity implements HasSupportFragme
                 return true;
             }
             case R.id.action_open_in_web: {
-                LinkHelper.openLink(getIntent().getStringExtra("url"), this);
+                LinkHelper.openLink(getIntent().getStringExtra(URL_EXTRA), this);
+                return true;
+            }
+            case R.id.action_reveal: {
+                fragment.onRevealPressed();
                 return true;
             }
         }
@@ -84,4 +131,5 @@ public class ViewThreadActivity extends BaseActivity implements HasSupportFragme
     public AndroidInjector<Fragment> supportFragmentInjector() {
         return dispatchingAndroidInjector;
     }
+
 }
