@@ -68,7 +68,7 @@ object AppInjector {
         if (activity is FragmentActivity) {
             activity.supportFragmentManager.registerFragmentLifecycleCallbacks(
                     object : FragmentManager.FragmentLifecycleCallbacks() {
-                        override fun onFragmentCreated(fm: FragmentManager?, f: Fragment?,
+                        override fun onFragmentCreated(fm: FragmentManager, f: Fragment,
                                                        savedInstanceState: Bundle?) {
                             if (f is Injectable) {
                                 AndroidSupportInjection.inject(f)
