@@ -16,11 +16,12 @@
 
 package com.keylesspalace.tusky.di
 
-import android.content.SharedPreferences
+import android.content.Context
 import android.text.Spanned
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonDeserializer
+import com.keylesspalace.tusky.BuildConfig
 import com.keylesspalace.tusky.db.AccountManager
 import com.keylesspalace.tusky.json.SpannedTypeAdapter
 import com.keylesspalace.tusky.network.InstanceSwitchAuthInterceptor
@@ -31,10 +32,11 @@ import dagger.Provides
 import dagger.multibindings.ClassKey
 import dagger.multibindings.IntoMap
 import dagger.multibindings.IntoSet
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Converter
 import retrofit2.Retrofit
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
 
@@ -67,32 +69,24 @@ class NetworkModule {
     fun providesConverterFactory(gson: Gson): Converter.Factory = GsonConverterFactory.create(gson)
 
     @Provides
-    @IntoSet
     @Singleton
-    fun providesAuthInterceptor(accountManager: AccountManager): Interceptor {
-        // should accept AccountManager here probably but I don't want to break things yet
-        return InstanceSwitchAuthInterceptor(accountManager)
-    }
-
-    @Provides
-    @Singleton
-    fun providesHttpClient(interceptors: @JvmSuppressWildcards Set<Interceptor>,
-                           preferences: SharedPreferences): OkHttpClient {
-        return OkHttpUtils.getCompatibleClientBuilder(preferences)
+    fun providesHttpClient(accountManager: AccountManager,
+                           context: Context): OkHttpClient {
+        return OkHttpUtils.getCompatibleClientBuilder(context)
                 .apply {
-                    interceptors.fold(this) { b, i ->
-                        b.addInterceptor(i)
+                    addInterceptor(InstanceSwitchAuthInterceptor(accountManager))
+                    if (BuildConfig.DEBUG) {
+                        addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BASIC))
                     }
                 }
                 .build()
     }
 
-
     @Provides
     @Singleton
     fun providesRetrofit(httpClient: OkHttpClient,
                          converters: @JvmSuppressWildcards Set<Converter.Factory>): Retrofit {
-        return Retrofit.Builder().baseUrl("https://dummy.placeholder/")
+        return Retrofit.Builder().baseUrl("https://"+MastodonApi.PLACEHOLDER_DOMAIN)
                 .client(httpClient)
                 .let { builder ->
                     // Doing it this way in case builder will be immutable so we return the final
@@ -100,6 +94,7 @@ class NetworkModule {
                     converters.fold(builder) { b, c ->
                         b.addConverterFactory(c)
                     }
+                    builder.addCallAdapterFactory(RxJava2CallAdapterFactory.createAsync())
                 }
                 .build()
 
