@@ -38,7 +38,13 @@ data class Account(
         val locked: Boolean = false,
         @SerializedName("followers_count") val followersCount: Int,
         @SerializedName("following_count") val followingCount: Int,
-        @SerializedName("statuses_count") val statusesCount: Int
+        @SerializedName("statuses_count") val statusesCount: Int,
+        val source: AccountSource?,
+        val bot: Boolean,
+        val emojis: List<Emoji>?,  // nullable for backward compatibility
+        val fields: List<Field>?,  //nullable for backward compatibility
+        val moved: Account? = null
+
 ) : Parcelable {
 
     val name: String
@@ -58,12 +64,33 @@ data class Account(
         return account?.id == this.id
     }
 
-    object SpannedParceler : Parceler<Spanned> {
-        override fun create(parcel: Parcel) = HtmlUtils.fromHtml(parcel.readString())
+    fun isRemote(): Boolean = this.username != this.localUsername
+}
 
-        override fun Spanned.write(parcel: Parcel, flags: Int) {
-            parcel.writeString(HtmlUtils.toHtml(this))
-        }
+@Parcelize
+data class AccountSource(
+        val privacy: Status.Visibility,
+        val sensitive: Boolean,
+        val note: String,
+        val fields: List<StringField>?
+): Parcelable
+
+@Parcelize
+data class Field (
+        val name: String,
+        val value: @WriteWith<SpannedParceler>() Spanned
+): Parcelable
+
+@Parcelize
+data class StringField (
+        val name: String,
+        val value: String
+): Parcelable
+
+object SpannedParceler : Parceler<Spanned> {
+    override fun create(parcel: Parcel): Spanned = HtmlUtils.fromHtml(parcel.readString())
+
+    override fun Spanned.write(parcel: Parcel, flags: Int) {
+        parcel.writeString(HtmlUtils.toHtml(this))
     }
-
 }
