@@ -15,14 +15,22 @@
 
 package com.keylesspalace.tusky.util;
 
+import android.content.ContentResolver;
+import android.net.Uri;
 import android.support.annotation.Nullable;
 
+import java.io.Closeable;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 
 public class IOUtils {
-    public static void closeQuietly(@Nullable InputStream stream) {
+
+    private static final int DEFAULT_BLOCKSIZE = 16384;
+
+    public static void closeQuietly(@Nullable Closeable stream) {
         try {
             if (stream != null) {
                 stream.close();
@@ -32,13 +40,32 @@ public class IOUtils {
         }
     }
 
-    public static void closeQuietly(@Nullable OutputStream stream) {
+    public static boolean copyToFile(ContentResolver contentResolver, Uri uri, File file) {
+        InputStream from;
+        FileOutputStream to;
         try {
-            if (stream != null) {
-                stream.close();
+            from = contentResolver.openInputStream(uri);
+            to = new FileOutputStream(file);
+        } catch (FileNotFoundException e) {
+            return false;
+        }
+        if (from == null) {
+            return false;
+        }
+        byte[] chunk = new byte[DEFAULT_BLOCKSIZE];
+        try {
+            while (true) {
+                int bytes = from.read(chunk, 0, chunk.length);
+                if (bytes < 0) {
+                    break;
+                }
+                to.write(chunk, 0, bytes);
             }
         } catch (IOException e) {
-            // intentionally unhandled
+            return false;
         }
+        closeQuietly(from);
+        closeQuietly(to);
+        return true;
     }
 }
