@@ -1,4 +1,5 @@
-/* Copyright 2017 Andrew Dawson
+/* Copyright 2018 Jeremiasz Nelz <remi6397(a)gmail.com>
+ * Copyright 2017 Andrew Dawson
  *
  * This file is a part of Tusky.
  *
@@ -23,21 +24,28 @@ import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.graphics.Color;
 import android.os.Build;
 import android.provider.Settings;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
+import android.support.v4.app.NotificationManagerCompat;
+import android.support.v4.app.RemoteInput;
 import android.support.v4.app.TaskStackBuilder;
 import android.support.v4.content.ContextCompat;
+import android.support.v4.text.BidiFormatter;
 import android.util.Log;
 
+import com.keylesspalace.tusky.BuildConfig;
 import com.keylesspalace.tusky.MainActivity;
 import com.keylesspalace.tusky.R;
-import com.keylesspalace.tusky.TuskyApplication;
 import com.keylesspalace.tusky.db.AccountEntity;
 import com.keylesspalace.tusky.db.AccountManager;
 import com.keylesspalace.tusky.entity.Notification;
+import com.keylesspalace.tusky.entity.Status;
 import com.keylesspalace.tusky.receiver.NotificationClearBroadcastReceiver;
+import com.keylesspalace.tusky.receiver.SendStatusBroadcastReceiver;
 import com.keylesspalace.tusky.view.RoundedTransformation;
 import com.squareup.picasso.Picasso;
 
@@ -46,9 +54,13 @@ import org.json.JSONException;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedHashSet;
 import java.util.List;
 
 public class NotificationHelper {
+
+    private static int notificationId = 0;
 
     /**
      * constants used in Intents
@@ -57,13 +69,39 @@ public class NotificationHelper {
 
     private static final String TAG = "NotificationHelper";
 
+    public static final String REPLY_ACTION = "REPLY_ACTION";
+
+    public static final String COMPOSE_ACTION = "COMPOSE_ACTION";
+
+    public static final String KEY_REPLY = "KEY_REPLY";
+
+    public static final String KEY_SENDER_ACCOUNT_ID = "KEY_SENDER_ACCOUNT_ID";
+
+    public static final String KEY_SENDER_ACCOUNT_IDENTIFIER = "KEY_SENDER_ACCOUNT_IDENTIFIER";
+
+    public static final String KEY_SENDER_ACCOUNT_FULL_NAME = "KEY_SENDER_ACCOUNT_FULL_NAME";
+
+    public static final String KEY_NOTIFICATION_ID = "KEY_NOTIFICATION_ID";
+
+    public static final String KEY_CITED_STATUS_ID = "KEY_CITED_STATUS_ID";
+
+    public static final String KEY_VISIBILITY = "KEY_VISIBILITY";
+
+    public static final String KEY_SPOILER = "KEY_SPOILER";
+
+    public static final String KEY_MENTIONS = "KEY_MENTIONS";
+
+    public static final String KEY_CITED_TEXT = "KEY_CITED_TEXT";
+
+    public static final String KEY_CITED_AUTHOR_LOCAL = "KEY_CITED_AUTHOR_LOCAL";
+
     /**
      * notification channels used on Android O+
      **/
-    private static final String CHANNEL_MENTION = "CHANNEL_MENTION";
-    private static final String CHANNEL_FOLLOW = "CHANNEL_FOLLOW";
-    private static final String CHANNEL_BOOST = "CHANNEL_BOOST";
-    private static final String CHANNEL_FAVOURITE = " CHANNEL_FAVOURITE";
+    public static final String CHANNEL_MENTION = "CHANNEL_MENTION";
+    public static final String CHANNEL_FOLLOW = "CHANNEL_FOLLOW";
+    public static final String CHANNEL_BOOST = "CHANNEL_BOOST";
+    public static final String CHANNEL_FAVOURITE = "CHANNEL_FAVOURITE";
 
     /**
      * Takes a given Mastodon notification and either creates a new Android notification or updates
@@ -74,7 +112,7 @@ public class NotificationHelper {
      * @param account the account for which the notification should be shown
      */
 
-    public static void make(final Context context, Notification body, AccountEntity account) {
+    public static void make(final Context context, Notification body, AccountEntity account, boolean isFirstOfBatch) {
 
         if (!filterNotification(account, body, context)) {
             return;
@@ -82,6 +120,7 @@ public class NotificationHelper {
 
         String rawCurrentNotifications = account.getActiveNotifications();
         JSONArray currentNotifications;
+        BidiFormatter bidiFormatter = BidiFormatter.getInstance();
 
         try {
             currentNotifications = new JSONArray(rawCurrentNotifications);
@@ -89,96 +128,195 @@ public class NotificationHelper {
             currentNotifications = new JSONArray();
         }
 
-        boolean alreadyContains = false;
-
         for (int i = 0; i < currentNotifications.length(); i++) {
             try {
                 if (currentNotifications.getString(i).equals(body.getAccount().getName())) {
-                    alreadyContains = true;
+                    currentNotifications.remove(i);
+                    break;
                 }
             } catch (JSONException e) {
                 Log.d(TAG, Log.getStackTraceString(e));
             }
         }
 
-        if (!alreadyContains) {
-            currentNotifications.put(body.getAccount().getName());
-        }
+        currentNotifications.put(body.getAccount().getName());
 
         account.setActiveNotifications(currentNotifications.toString());
 
-        //no need to save account, this will be done in the calling function
+        // Notification group member
+        // =========================
+        final NotificationCompat.Builder builder = newNotification(context, body, account, false);
 
-        Intent resultIntent = new Intent(context, MainActivity.class);
-        resultIntent.putExtra(ACCOUNT_ID, account.getId());
-        TaskStackBuilder stackBuilder = TaskStackBuilder.create(context);
-        stackBuilder.addParentStack(MainActivity.class);
-        stackBuilder.addNextIntent(resultIntent);
-        PendingIntent resultPendingIntent = stackBuilder.getPendingIntent((int) account.getId(),
-                PendingIntent.FLAG_UPDATE_CURRENT);
+        notificationId++;
 
-        Intent deleteIntent = new Intent(context, NotificationClearBroadcastReceiver.class);
-        deleteIntent.putExtra(ACCOUNT_ID, account.getId());
-        PendingIntent deletePendingIntent = PendingIntent.getBroadcast(context, (int) account.getId(), deleteIntent,
-                PendingIntent.FLAG_UPDATE_CURRENT);
+        builder.setContentTitle(titleForType(context, body, bidiFormatter))
+                .setContentText(bodyForType(body));
 
-        final NotificationCompat.Builder builder = new NotificationCompat.Builder(context, getChannelId(account, body))
-                .setSmallIcon(R.drawable.ic_notify)
-                .setContentIntent(resultPendingIntent)
-                .setDeleteIntent(deletePendingIntent)
-                .setColor(ContextCompat.getColor(context, (R.color.primary)))
-                .setDefaults(0); // So it doesn't ring twice, notify only in Target callback
+        if (body.getType() == Notification.Type.MENTION) {
+            builder.setStyle(new NotificationCompat.BigTextStyle()
+                    .bigText(bodyForType(body)));
+        }
 
-        setupPreferences(account, builder);
+        //load the avatar synchronously
+        Bitmap accountAvatar;
+        try {
+            accountAvatar = Picasso.with(context)
+                    .load(body.getAccount().getAvatar())
+                    .transform(new RoundedTransformation(20))
+                    .get();
+        } catch (IOException e) {
+            Log.d(TAG, "error loading account avatar", e);
+            accountAvatar = BitmapFactory.decodeResource(context.getResources(), R.drawable.avatar_default);
+        }
 
-        if (currentNotifications.length() == 1) {
-            builder.setContentTitle(titleForType(context, body))
-                    .setContentText(bodyForType(body));
+        builder.setLargeIcon(accountAvatar);
 
-            if (body.getType() == Notification.Type.MENTION) {
-                builder.setStyle(new NotificationCompat.BigTextStyle()
-                        .bigText(bodyForType(body)));
-            }
+        // Reply to mention action; RemoteInput is available from KitKat Watch, but buttons are available from Nougat
+        if (body.getType() == Notification.Type.MENTION
+                && android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            RemoteInput replyRemoteInput = new RemoteInput.Builder(KEY_REPLY)
+                    .setLabel(context.getString(R.string.label_quick_reply))
+                    .build();
 
-            //load the avatar synchronously
-            Bitmap accountAvatar;
-            try {
-                accountAvatar = Picasso.with(context)
-                        .load(body.getAccount().getAvatar())
-                        .transform(new RoundedTransformation(20))
-                        .get();
-            } catch (IOException e) {
-                Log.d(TAG, "error loading account avatar", e);
-                accountAvatar = BitmapFactory.decodeResource(context.getResources(), R.drawable.avatar_default);
-            }
+            PendingIntent quickReplyPendingIntent = getStatusReplyIntent(REPLY_ACTION, context, body, account);
 
-            builder.setLargeIcon(accountAvatar);
+            NotificationCompat.Action quickReplyAction =
+                    new NotificationCompat.Action.Builder(R.drawable.ic_reply_24dp,
+                            context.getString(R.string.action_quick_reply), quickReplyPendingIntent)
+                            .addRemoteInput(replyRemoteInput)
+                            .build();
 
-        } else {
+            builder.addAction(quickReplyAction);
+
+            PendingIntent composePendingIntent = getStatusReplyIntent(COMPOSE_ACTION, context, body, account);
+
+            NotificationCompat.Action composeAction =
+                    new NotificationCompat.Action.Builder(R.drawable.ic_reply_24dp,
+                            context.getString(R.string.action_compose_shortcut), composePendingIntent)
+                            .build();
+
+            builder.addAction(composeAction);
+        }
+
+        builder.setSubText(account.getFullName());
+        builder.setVisibility(NotificationCompat.VISIBILITY_PRIVATE);
+        builder.setCategory(NotificationCompat.CATEGORY_SOCIAL);
+        builder.setOnlyAlertOnce(true);
+
+        // only alert for the first notification of a batch to avoid multiple alerts at once
+        if(!isFirstOfBatch) {
+            builder.setGroupAlertBehavior(NotificationCompat.GROUP_ALERT_SUMMARY);
+        }
+
+        // Summary
+        // =======
+        final NotificationCompat.Builder summaryBuilder = newNotification(context, body, account, true);
+
+        if (currentNotifications.length() != 1) {
             try {
                 String title = context.getString(R.string.notification_title_summary, currentNotifications.length());
-                String text = joinNames(context, currentNotifications);
-                builder.setContentTitle(title)
+                String text = joinNames(context, currentNotifications, bidiFormatter);
+                summaryBuilder.setContentTitle(title)
                         .setContentText(text);
             } catch (JSONException e) {
                 Log.d(TAG, Log.getStackTraceString(e));
             }
         }
 
-        builder.setSubText(account.getFullName());
+        summaryBuilder.setSubText(account.getFullName());
+        summaryBuilder.setVisibility(NotificationCompat.VISIBILITY_PRIVATE);
+        summaryBuilder.setCategory(NotificationCompat.CATEGORY_SOCIAL);
+        summaryBuilder.setOnlyAlertOnce(true);
+        summaryBuilder.setGroupSummary(true);
 
-        builder.setVisibility(NotificationCompat.VISIBILITY_PRIVATE);
-        builder.setCategory(NotificationCompat.CATEGORY_SOCIAL);
-
-        builder.setOnlyAlertOnce(true);
-
-        NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(context);
 
         //noinspection ConstantConditions
-        notificationManager.notify((int) account.getId(), builder.build());
+        notificationManager.notify(notificationId, builder.build());
+        if (currentNotifications.length() == 1) {
+            notificationManager.notify((int) account.getId(), builder.setGroupSummary(true).build());
+        } else {
+            notificationManager.notify((int) account.getId(), summaryBuilder.build());
+        }
     }
 
-    public static void createNotificationChannelsForAccount(AccountEntity account, Context context) {
+    private static NotificationCompat.Builder newNotification(Context context, Notification body, AccountEntity account, boolean summary) {
+        Intent summaryResultIntent = new Intent(context, MainActivity.class);
+        summaryResultIntent.putExtra(ACCOUNT_ID, account.getId());
+        TaskStackBuilder summaryStackBuilder = TaskStackBuilder.create(context);
+        summaryStackBuilder.addParentStack(MainActivity.class);
+        summaryStackBuilder.addNextIntent(summaryResultIntent);
+
+        PendingIntent summaryResultPendingIntent = summaryStackBuilder.getPendingIntent(notificationId,
+                PendingIntent.FLAG_UPDATE_CURRENT);
+
+        // we have to switch account here
+        Intent eventResultIntent = new Intent(context, MainActivity.class);
+        eventResultIntent.putExtra(ACCOUNT_ID, account.getId());
+        TaskStackBuilder eventStackBuilder = TaskStackBuilder.create(context);
+        eventStackBuilder.addParentStack(MainActivity.class);
+        eventStackBuilder.addNextIntent(eventResultIntent);
+
+        PendingIntent eventResultPendingIntent = eventStackBuilder.getPendingIntent((int) account.getId(),
+                PendingIntent.FLAG_UPDATE_CURRENT);
+
+        Intent deleteIntent = new Intent(context, NotificationClearBroadcastReceiver.class);
+        deleteIntent.putExtra(ACCOUNT_ID, account.getId());
+        PendingIntent deletePendingIntent = PendingIntent.getBroadcast(context, summary ? (int) account.getId() : notificationId, deleteIntent,
+                PendingIntent.FLAG_UPDATE_CURRENT);
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(context, getChannelId(account, body))
+                .setSmallIcon(R.drawable.ic_notify)
+                .setContentIntent(summary ? summaryResultPendingIntent : eventResultPendingIntent)
+                .setDeleteIntent(deletePendingIntent)
+                .setColor(BuildConfig.DEBUG ? Color.parseColor("#19A341") : ContextCompat.getColor(context, R.color.primary))
+                .setGroup(account.getAccountId())
+                .setAutoCancel(true)
+                .setDefaults(0); // So it doesn't ring twice, notify only in Target callback
+
+        setupPreferences(account, builder);
+
+        return builder;
+    }
+
+    private static PendingIntent getStatusReplyIntent(String action, Context context, Notification body, AccountEntity account) {
+        Status status = body.getStatus();
+
+        String citedLocalAuthor = status.getAccount().getLocalUsername();
+        String citedText = status.getContent().toString();
+        String inReplyToId = status.getId();
+        Status actionableStatus = status.getActionableStatus();
+        Status.Visibility replyVisibility = actionableStatus.getVisibility();
+        String contentWarning = actionableStatus.getSpoilerText();
+        Status.Mention[] mentions = actionableStatus.getMentions();
+        List<String> mentionedUsernames = new ArrayList<>();
+        mentionedUsernames.add(actionableStatus.getAccount().getUsername());
+        for (Status.Mention mention : mentions) {
+            mentionedUsernames.add(mention.getUsername());
+        }
+        mentionedUsernames.removeAll(Collections.singleton(account.getUsername()));
+        mentionedUsernames = new ArrayList<>(new LinkedHashSet<>(mentionedUsernames));
+
+        Intent replyIntent = new Intent(context, SendStatusBroadcastReceiver.class)
+                .setAction(action)
+                .putExtra(KEY_CITED_AUTHOR_LOCAL, citedLocalAuthor)
+                .putExtra(KEY_CITED_TEXT, citedText)
+                .putExtra(KEY_SENDER_ACCOUNT_ID, account.getId())
+                .putExtra(KEY_SENDER_ACCOUNT_IDENTIFIER, account.getIdentifier())
+                .putExtra(KEY_SENDER_ACCOUNT_FULL_NAME, account.getFullName())
+                .putExtra(KEY_NOTIFICATION_ID, notificationId)
+                .putExtra(KEY_CITED_STATUS_ID, inReplyToId)
+                .putExtra(KEY_VISIBILITY, replyVisibility)
+                .putExtra(KEY_SPOILER, contentWarning)
+                .putExtra(KEY_MENTIONS, mentionedUsernames.toArray(new String[0]));
+
+        return PendingIntent.getBroadcast(context.getApplicationContext(),
+                notificationId,
+                replyIntent,
+                PendingIntent.FLAG_UPDATE_CURRENT);
+    }
+
+    public static void createNotificationChannelsForAccount(@NonNull AccountEntity account, @NonNull Context context) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
 
             NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
@@ -230,7 +368,7 @@ public class NotificationHelper {
         }
     }
 
-    public static void deleteNotificationChannelsForAccount(AccountEntity account, Context context) {
+    public static void deleteNotificationChannelsForAccount(@NonNull AccountEntity account, @NonNull Context context) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
 
             NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
@@ -241,21 +379,26 @@ public class NotificationHelper {
         }
     }
 
-    public static void deleteLegacyNotificationChannels(Context context) {
-        // delete the notification channels that where used before the multi account mode was introduced to avoid confusion
+    public static void deleteLegacyNotificationChannels(@NonNull Context context, @NonNull AccountManager accountManager) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
 
             NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
 
+            // used until Tusky 1.4
             //noinspection ConstantConditions
             notificationManager.deleteNotificationChannel(CHANNEL_MENTION);
             notificationManager.deleteNotificationChannel(CHANNEL_FAVOURITE);
             notificationManager.deleteNotificationChannel(CHANNEL_BOOST);
             notificationManager.deleteNotificationChannel(CHANNEL_FOLLOW);
+
+            // used until Tusky 1.7
+            for(AccountEntity account: accountManager.getAllAccountsOrderedByActive()) {
+                notificationManager.deleteNotificationChannel(CHANNEL_FAVOURITE+" "+account.getIdentifier());
+            }
         }
     }
 
-    public static boolean areNotificationsEnabled(Context context) {
+    public static boolean areNotificationsEnabled(@NonNull Context context, @NonNull AccountManager accountManager) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
 
             // on Android >= O, notifications are enabled, if at least one channel is enabled
@@ -276,15 +419,12 @@ public class NotificationHelper {
 
         } else {
             // on Android < O, notifications are enabled, if at least one account has notification enabled
-            return TuskyApplication.getInstance(context).getServiceLocator()
-                    .get(AccountManager.class).areNotificationsEnabled();
+            return accountManager.areNotificationsEnabled();
         }
 
     }
 
-    public static void clearNotificationsForActiveAccount(Context context) {
-        AccountManager accountManager = TuskyApplication.getInstance(context).getServiceLocator()
-                .get(AccountManager.class);
+    public static void clearNotificationsForActiveAccount(@NonNull Context context, @NonNull AccountManager accountManager) {
         AccountEntity account = accountManager.getActiveAccount();
         if (account != null) {
             account.setActiveNotifications("[]");
@@ -355,43 +495,53 @@ public class NotificationHelper {
         }
     }
 
+    private static String wrapItemAt(JSONArray array, int index, BidiFormatter bidiFormatter) throws JSONException {
+        return bidiFormatter.unicodeWrap(array.get(index).toString());
+    }
+
     @Nullable
-    private static String joinNames(Context context, JSONArray array) throws JSONException {
+    private static String joinNames(Context context, JSONArray array, BidiFormatter bidiFormatter) throws JSONException {
         if (array.length() > 3) {
             int length = array.length();
             return String.format(context.getString(R.string.notification_summary_large),
-                    array.get(length - 1), array.get(length - 2), array.get(length - 3), length - 3);
+                    wrapItemAt(array, length - 1, bidiFormatter),
+                    wrapItemAt(array, length - 2, bidiFormatter),
+                    wrapItemAt(array, length - 3, bidiFormatter),
+                    length - 3);
         } else if (array.length() == 3) {
             return String.format(context.getString(R.string.notification_summary_medium),
-                    array.get(2), array.get(1), array.get(0));
+                    wrapItemAt(array, 2, bidiFormatter),
+                    wrapItemAt(array, 1, bidiFormatter),
+                    wrapItemAt(array, 0, bidiFormatter));
         } else if (array.length() == 2) {
             return String.format(context.getString(R.string.notification_summary_small),
-                    array.get(1), array.get(0));
+                    wrapItemAt(array, 1, bidiFormatter),
+                    wrapItemAt(array, 0, bidiFormatter));
         }
 
         return null;
     }
 
     @Nullable
-    private static String titleForType(Context context, Notification notification) {
+    private static String titleForType(Context context, Notification notification, BidiFormatter bidiFormatter) {
+        String accountName = bidiFormatter.unicodeWrap(notification.getAccount().getName());
         switch (notification.getType()) {
             case MENTION:
                 return String.format(context.getString(R.string.notification_mention_format),
-                        notification.getAccount().getName());
+                        accountName);
             case FOLLOW:
                 return String.format(context.getString(R.string.notification_follow_format),
-                        notification.getAccount().getName());
+                        accountName);
             case FAVOURITE:
                 return String.format(context.getString(R.string.notification_favourite_format),
-                        notification.getAccount().getName());
+                        accountName);
             case REBLOG:
                 return String.format(context.getString(R.string.notification_reblog_format),
-                        notification.getAccount().getName());
+                        accountName);
         }
         return null;
     }
 
-    @Nullable
     private static String bodyForType(Notification notification) {
         switch (notification.getType()) {
             case FOLLOW:
@@ -399,7 +549,11 @@ public class NotificationHelper {
             case MENTION:
             case FAVOURITE:
             case REBLOG:
-                return notification.getStatus().getContent().toString();
+                if (notification.getStatus().getSensitive()) {
+                    return notification.getStatus().getSpoilerText();
+                } else {
+                    return notification.getStatus().getContent().toString();
+                }
         }
         return null;
     }
