@@ -30,15 +30,16 @@ import android.support.v7.app.AppCompatDelegate;
 import android.util.TypedValue;
 import android.widget.ImageView;
 
-import com.keylesspalace.tusky.TuskyApplication;
-
 /**
  * Provides runtime compatibility to obtain theme information and re-theme views, especially where
  * the ability to do so is not supported in resource files.
  */
 public class ThemeUtils {
+    public static final String APP_THEME_DEFAULT = ThemeUtils.THEME_NIGHT;
+
     public static final String THEME_NIGHT = "night";
     public static final String THEME_DAY = "day";
+    public static final String THEME_BLACK = "black";
     public static final String THEME_AUTO = "auto";
 
     public static Drawable getDrawable(Context context, @AttrRes int attribute,
@@ -91,7 +92,7 @@ public class ThemeUtils {
         drawable.setColorFilter(getColor(context, attribute), PorterDuff.Mode.SRC_IN);
     }
 
-    public static void setAppNightMode(String flavor) {
+    public static void setAppNightMode(String flavor, Context context) {
         int mode;
         switch (flavor) {
             default:
@@ -101,13 +102,17 @@ public class ThemeUtils {
             case THEME_DAY:
                 mode = UiModeManager.MODE_NIGHT_NO;
                 break;
+            case THEME_BLACK:
+                mode = UiModeManager.MODE_NIGHT_YES;
+                break;
             case THEME_AUTO:
                 mode = UiModeManager.MODE_NIGHT_AUTO;
                 break;
         }
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            TuskyApplication.getUiModeManager().setNightMode(mode);
+            UiModeManager uiModeManager = (UiModeManager)context.getApplicationContext().getSystemService(Context.UI_MODE_SERVICE);
+            uiModeManager.setNightMode(mode);
         } else {
             AppCompatDelegate.setDefaultNightMode(mode);
         }
