@@ -16,20 +16,33 @@
 
 package com.keylesspalace.tusky
 
+import android.text.SpannedString
 import android.widget.EditText
 import com.keylesspalace.tusky.db.AccountEntity
 import com.keylesspalace.tusky.db.AccountManager
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
+import com.keylesspalace.tusky.db.AppDatabase
+import com.keylesspalace.tusky.db.InstanceDao
+import com.keylesspalace.tusky.entity.Account
+import com.keylesspalace.tusky.entity.Emoji
+import com.keylesspalace.tusky.entity.Instance
+import com.keylesspalace.tusky.network.MastodonApi
+import okhttp3.Request
+import okhttp3.ResponseBody
+import org.junit.Assert
+import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito
 import org.mockito.Mockito.`when`
+import org.mockito.Mockito.mock
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import org.robolectric.fakes.RoboMenuItem
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 /**
  * Created by charlag on 3/7/18.
@@ -40,9 +53,8 @@ import org.robolectric.fakes.RoboMenuItem
 class ComposeActivityTest {
 
     lateinit var activity: ComposeActivity
-    lateinit var application: FakeTuskyApplication
-    lateinit var serviceLocator: TuskyApplication.ServiceLocator
     lateinit var accountManagerMock: AccountManager
+    lateinit var apiMock: MastodonApi
 
     val account = AccountEntity(
             id = 1,
@@ -62,18 +74,74 @@ class ComposeActivityTest {
             notificationVibration = true,
             notificationLight = true
     )
+    var instanceResponseCallback: ((Call<Instance>?, Callback<Instance>?)->Unit)? = null
 
     @Before
-    fun before() {
+    fun setupActivity() {
         val controller = Robolectric.buildActivity(ComposeActivity::class.java)
         activity = controller.get()
+
         accountManagerMock = Mockito.mock(AccountManager::class.java)
-        serviceLocator = Mockito.mock(TuskyApplication.ServiceLocator::class.java)
-        `when`(serviceLocator.get(AccountManager::class.java)).thenReturn(accountManagerMock)
-        `when`(accountManagerMock.activeAccount).thenReturn(account)
+
+        apiMock = Mockito.mock(MastodonApi::class.java)
+        `when`(apiMock.customEmojis).thenReturn(object: Call<List<Emoji>> {
+            override fun isExecuted(): Boolean {
+                return false
+            }
+            override fun clone(): Call<List<Emoji>> {
+                throw Error("not implemented")
+            }
+            override fun isCanceled(): Boolean {
+                throw Error("not implemented")
+            }
+            override fun cancel() {
+                throw Error("not implemented")
+            }
+            override fun execute(): Response<List<Emoji>> {
+                throw Error("not implemented")
+            }
+            override fun request(): Request {
+                throw Error("not implemented")
+            }
+
+            override fun enqueue(callback: Callback<List<Emoji>>?) {}
+        })
+        `when`(apiMock.instance).thenReturn(object: Call<Instance> {
+            override fun isExecuted(): Boolean {
+                return false
+            }
+            override fun clone(): Call<Instance> {
+                throw Error("not implemented")
+            }
+            override fun isCanceled(): Boolean {
+                throw Error("not implemented")
+            }
+            override fun cancel() {
+                throw Error("not implemented")
+            }
+            override fun execute(): Response<Instance> {
+                throw Error("not implemented")
+            }
+            override fun request(): Request {
+                throw Error("not implemented")
+            }
+
+            override fun enqueue(callback: Callback<Instance>?) {
+                instanceResponseCallback?.invoke(this, callback)
+            }
+        })
+
+        val instanceDaoMock = mock(InstanceDao::class.java)
+        val dbMock = mock(AppDatabase::class.java)
+        `when`(dbMock.instanceDao()).thenReturn(instanceDaoMock)
+
+        activity.mastodonApi = apiMock
         activity.accountManager = accountManagerMock
-        application = activity.application as FakeTuskyApplication
-        application.locator = serviceLocator
+        activity.database = dbMock
+
+        `when`(accountManagerMock.activeAccount).thenReturn(account)
+
+
         controller.create().start()
     }
 
@@ -105,6 +173,65 @@ class ComposeActivityTest {
         // We would like to check for dialog but Robolectric doesn't work with AppCompat v7 yet
     }
 
+    @Test
+    fun whenMaximumTootCharsIsNull_defaultLimitIsUsed() {
+        instanceResponseCallback = getSuccessResponseCallbackWithMaximumTootCharacters(null)
+        setupActivity()
+        assertEquals(ComposeActivity.STATUS_CHARACTER_LIMIT, activity.maximumTootCharacters)
+    }
+
+    @Test
+    fun whenMaximumTootCharsIsPopulated_customLimitIsUsed() {
+        val customMaximum = 1000
+        instanceResponseCallback = getSuccessResponseCallbackWithMaximumTootCharacters(customMaximum)
+        setupActivity()
+        assertEquals(customMaximum, activity.maximumTootCharacters)
+    }
+
+    @Test
+    fun whenInitialInstanceRequestFails_defaultValueIsUsed() {
+        instanceResponseCallback = {
+            call: Call<Instance>?, callback: Callback<Instance>? ->
+            if (call != null) {
+                callback?.onResponse(call, Response.error(400, ResponseBody.create(null, "")))
+            }
+        }
+        setupActivity()
+        assertEquals(ComposeActivity.STATUS_CHARACTER_LIMIT, activity.maximumTootCharacters)
+    }
+
+    @Test
+    fun whenTextContainsNoUrl_everyCharacterIsCounted() {
+        val content = "This is test content please ignore thx "
+        insertSomeTextInContent(content)
+        Assert.assertEquals(activity.calculateTextLength(), content.length)
+    }
+
+    @Test
+    fun whenTextContainsUrl_onlyEllipsizedURLIsCounted() {
+        val url = "https://www.google.dk/search?biw=1920&bih=990&tbm=isch&sa=1&ei=bmDrWuOoKMv6kwWOkIaoDQ&q=indiana+jones+i+hate+snakes+animated&oq=indiana+jones+i+hate+snakes+animated&gs_l=psy-ab.3...54174.55443.0.55553.9.7.0.0.0.0.255.333.1j0j1.2.0....0...1c.1.64.psy-ab..7.0.0....0.40G-kcDkC6A#imgdii=PSp15hQjN1JqvM:&imgrc=H0hyE2JW5wrpBM:"
+        val additionalContent = "Check out this @image #search result: "
+        insertSomeTextInContent(additionalContent + url)
+        Assert.assertEquals(activity.calculateTextLength(), additionalContent.length + ComposeActivity.MAXIMUM_URL_LENGTH)
+    }
+
+    @Test
+    fun whenTextContainsMultipleUrls_onlyEllipsizedURLIsCounted() {
+        val shortUrl = "https://tusky.app"
+        val url = "https://www.google.dk/search?biw=1920&bih=990&tbm=isch&sa=1&ei=bmDrWuOoKMv6kwWOkIaoDQ&q=indiana+jones+i+hate+snakes+animated&oq=indiana+jones+i+hate+snakes+animated&gs_l=psy-ab.3...54174.55443.0.55553.9.7.0.0.0.0.255.333.1j0j1.2.0....0...1c.1.64.psy-ab..7.0.0....0.40G-kcDkC6A#imgdii=PSp15hQjN1JqvM:&imgrc=H0hyE2JW5wrpBM:"
+        val additionalContent = " Check out this @image #search result: "
+        insertSomeTextInContent(shortUrl + additionalContent + url)
+        Assert.assertEquals(activity.calculateTextLength(), additionalContent.length + shortUrl.length + ComposeActivity.MAXIMUM_URL_LENGTH)
+    }
+
+    @Test
+    fun whenTextContainsMultipleURLs_allURLsGetEllipsized() {
+        val url = "https://www.google.dk/search?biw=1920&bih=990&tbm=isch&sa=1&ei=bmDrWuOoKMv6kwWOkIaoDQ&q=indiana+jones+i+hate+snakes+animated&oq=indiana+jones+i+hate+snakes+animated&gs_l=psy-ab.3...54174.55443.0.55553.9.7.0.0.0.0.255.333.1j0j1.2.0....0...1c.1.64.psy-ab..7.0.0....0.40G-kcDkC6A#imgdii=PSp15hQjN1JqvM:&imgrc=H0hyE2JW5wrpBM:"
+        val additionalContent = " Check out this @image #search result: "
+        insertSomeTextInContent(url + additionalContent + url)
+        Assert.assertEquals(activity.calculateTextLength(), additionalContent.length + (ComposeActivity.MAXIMUM_URL_LENGTH * 2))
+    }
+
     private fun clickUp() {
         val menuItem = RoboMenuItem(android.R.id.home)
         activity.onOptionsItemSelected(menuItem)
@@ -114,7 +241,51 @@ class ComposeActivityTest {
         activity.onBackPressed()
     }
 
-    private fun insertSomeTextInContent() {
-        activity.findViewById<EditText>(R.id.compose_edit_field).setText("Some text")
+    private fun insertSomeTextInContent(text: String? = null) {
+        activity.findViewById<EditText>(R.id.composeEditField).setText(text ?: "Some text")
+    }
+
+    private fun getInstanceWithMaximumTootCharacters(maximumTootCharacters: Int?): Instance
+    {
+        return Instance(
+                "https://example.token",
+                "Example dot Token",
+                "Example instance for testing",
+                "admin@example.token",
+                "2.6.3",
+                HashMap<String, String>(),
+                null,
+                null,
+                listOf("en"),
+                Account(
+                        "1",
+                        "admin",
+                        "admin",
+                        "admin",
+                        SpannedString(""),
+                        "https://example.token",
+                        "",
+                        "",
+                        false,
+                        0,
+                        0,
+                        0,
+                        null,
+                        false,
+                        emptyList(),
+                        emptyList()
+                ),
+                maximumTootCharacters
+        )
+    }
+
+    private fun getSuccessResponseCallbackWithMaximumTootCharacters(maximumTootCharacters: Int?): (Call<Instance>?, Callback<Instance>?) -> Unit
+    {
+        return {
+            call: Call<Instance>?, callback: Callback<Instance>? ->
+            if (call != null) {
+                callback?.onResponse(call, Response.success(getInstanceWithMaximumTootCharacters(maximumTootCharacters)))
+            }
+        }
     }
 }
